requirement.txt
# requirements.txt
opencv-python
mediapipe
numpy
Pillow
websockets
